An MVP solution for displaying more error information in VenueDetailsForm
