+++ src/components/molecules/ImageInput/ImageCollectionInput.tsx
@@ -130,12 +130,7 @@ export const ImageCollectionInput: React.FC<ImageInputProps> = (props) => {
             )}
         </div>
       </div>
-      <input
-        type="hidden"
-        name={`${fieldName}Url`}
-        ref={register}
-        value={imageUrlForPreview}
-      />
+      <input type="hidden" name={`${fieldName}Url`} ref={register} />
       {error?.message && <span className="input-error">{error.message}</span>}
       {hasImageCollections && (
         <div style={{ marginTop: 10, fontSize: "16px" }}>
